requirements.txt
google-adk
streamlit
python-dotenv
pandas
plotly
pydantic
litellm
backoff
requests
